Refact: deprecated 된 스크롤 훅 코드
