+++ library/waffleScroll/types.ts
@@ -43,13 +43,6 @@ export type GlobalScrollHook<T extends Record<string, unknown>> = (
   targetRef: MutableRefObject<AvailableHTMLElement | null>;
   state: T;
 };
-export type LocalScrollHook<
-  T extends Record<string, unknown>,
-  U extends Record<string, unknown>,
-> = (params: { callback?: ScrollCallback<T & U>; initialState?: U }) => {
-  targetRef: MutableRefObject<AvailableHTMLElement | null>;
-  state: T;
-};
 
 export type SetScrollContainer = (
   containerElement: AvailableHTMLElement,
@@ -69,8 +62,3 @@ export type ScrollCreatorReturnType<T extends Record<string, unknown>> =
     setScrollContainer: SetScrollContainer;
     scrollTo: ScrollTo;
   } & ScrollApis<T>;
-
-export type LocalScrollCreatorReturnType<
-  T extends Record<string, unknown>,
-  U extends Record<string, unknown>,
-> = LocalScrollHook<T, U> & { setScrollContainer: SetScrollContainer };
